fota: introduce flash block buffer write support
SoCs can vary in flash write alignment settings.  To avoid most of
these issues, let's store the incoming buffer data into "blocks"
which are written to flash in a single command.

This also has the advantage of minimizing block writes which currently
happen for every netbuf fragment (74 to 128 bytes).

A "block" is defined as 512 bytes.  But in the future we can add the
ability to customize this value per board.

Signed-off-by: Michael Scott <michael.scott@linaro.org>