frdm-k64f: add off-chip temperature support
The accelerometer/magnetometer on the FRDM-K64F has an (uncalibrated)
temperature sensor as well. Use that as the "fota-offchip-temp"
device.

To make this happen, we need configuration that's only meaningful on
this board. Rather than have users specify multiple CONF_FILE values,
borrow a trick from the mcuboot playbook and find the right one based
on the BOARD setting.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>