consolidate hawkbit ACID code into hawkbit.c
We're currently pretending that different device management backends
will give us some sort of ID to put into the application state
partition to manage the result of a single OTA update.

That's really a hawkBit-specific assumption, though, so move that code
into hawkbit.c. While we're here, do some renaming so the code is
easier to understand.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>