hawkbit: use net mgmt API to make sure net inferface is ready
Instead of using a variable created in main.c to see whether bluetooth
is connected or not, let's make this generic and use net management APIs
to make sure whatever network interface we are using is ready before
sending data.

This removes the dependency to main.c's and it's bt_connection_state
variable (which was causing a build break).

Signed-off-by: Michael Scott <michael.scott@linaro.org>