Run an application work queue from the main thread.
This patch allows this application to publish test case results using
tc_util.h APIs, from multiple threads, and after the BIST is complete.

We want to use tc_util.h instead of sys_log.h to display test
results. This will define the line clearly between log messages, which
are read by humans and may change, and test reporting, which is also
meant for parsing by machines, and needs a fixed display format.

In our testing environment, all tc_util.h calls between TC_START() and
TC_END_REPORT() must belong to the same suite, or the parsers will
break. Since we need to publish results from multiple
threads (Bluemix, hawkBit), we need to schedule these calls
sequentially.

Existing Zephyr work queue APIs are not sufficient for this:

- Since tc_util.h calls block on the UART, we can't use
  k_work_submit(), etc.

- Using k_work_q_start() isn't practical, either. We're low on memory;
  allocating another stack would hurt.

What we'd really like is the ability to run a work queue from an
existing thread: the main thread isn't doing anything important after
the BIST is over, so it can handle this.

However, there's no Zephyr API for that.

Thus, add app_work_queue.h, which gives us what we want, and use it
when main() is done with the BIST. As a consequence, schedule LED
toggles using delayed work on this queue, instead of sleeping.

Future patches can use this API to schedule work to print test
results. There may be other uses as well.

For now, this copies the Zephyr work queue implementation into the
application; we'll need to think about how to upstream this.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>