dm-hawkbit-mqtt: report deployment feedback with zephyr JSON
Replace hand-rolled JSON for the deployment feedback resource with
descriptors and calls to the Zephyr JSON library.

Due to limitations in the Zephyr JSON API, we cannot define a data
schema and then encode buffers which have some of the fields
missing. For that reason, some unused fields are deliberately left
out of the definitions in hawkbit_priv.h.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>