bluetooth: fix BT MAC address for new kernels
Kernels newer than 4.11 require the removal of ZEP1656 config due
to several bluetooth / 6lowpan bugs being fixed in the mainline Linux
kernel.

One of these changes stops toggling the U/L bit of the BT MAC address
as a way of knowing whether the MAC is public or random.  Due to this
change we need to set octet 5 of the MAC correctly, as it will not be
adjusted when ZEP1656 is disabled.

This fixes communication problems with several nodes connected via
6lowpan on kernels > 4.11.

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>