dm-hawkbit-mqtt: get rid of most of "bluemix"
At heart, this is a generic MQTT application. The "bluemix" pieces are
leftovers from when we were trying to support that cloud backend. It's
time to remove them all.

Do a bunch of renaming to get rid of most of the Bluemix (and "bm")
references, but otherwise leave the structure of the program the same.

There are two places where it's not possible to do that without
changing the semantics in a way that feels a little too intense for a
find/replace patch:

1. the JSON status publisher (since the data schema is
   Bluemix-specific)

2. the contents of what is now struct mqtt_helper_ctx (which are used
   to log in to what this app still expects to be a Bluemix instance).

Leave those the way they are. A more invasive change in a subsequent
patch will merge src/lib/mqtt-helper.c into src/mqtt_temperature.c; we
can deal with those changes then.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>