dm-hawkbit-mqtt: remove lib/mqtt-helper
Now that it's clear that the "helper" is really just a bit of extra
and unnecessary threading and bluemix-related configuration on top of
a standard MQTT publisher, it's time to remove the extra "lib" files
related to MQTT.

The purpose of the lib/ directory is to contain extra library sources
that we'd like to use from multiple places. This file does not qualify.

To handle this, merge src/lib/mqtt-helper.c into
src/mqtt_temperature.c. The resulting code is also easier to read,
which is nice. Some strange names have been kept as-is to make review
easier. Further clean-up patches will make this better.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>