dm-hawkbit-mqtt: reorganize hawkbit internals
Consolidate state into the struct hawkbit_context. This will make it
easier to run hawkbit from the application work queue. In general
prefer to take a struct hawkbit_context* instead of a struct
hawkbit_download*.

Use a pointer named "hbc" consistently to refer to the hawkbit context
argument to each function, as opposed to sometimes referring to the
global and using various names.

Rename the semaphore in the hawkbit context to "sem", as it's
useful for more than just waiting on a download.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>