dm-hawkbit-mqtt: add extra packet pool needed by BT L2 for MQTT
6lowpan compresses network packets prior to sending "over the wire".
During retransmission processing, an original copy of the packet
needs to be held so that it can be re-compressed again.  The Zephyr
TCP stack makes use of the CONFIG_NET_CONTEXT_NET_PKT_POOL setting
to establish a "holding" pool for this purpose.

This should help with -ENOMEM errors showing up randomly during MQTT
retransissions.

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>