dm-hawkbit_mqtt: use json library for sensor data
Use the upstream Zephyr json.h interface when serializing sensor data
instead of creating it by hand.

This unfortunately consumes a bit of extra RAM to create a descriptors
array, since we find our devices dynamically. This could be fixed, and
the descriptor array be made a compile-time constant in flash, by
using device-specific device tree overlays to declare the sensors
present on each device.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>