Don't assume NET_L2_BT for BT-enabled HW
Let's remove the auto-selection of NET_L2_BT and move it to the
board/*.conf files which support bluetooth.

This way, an overlay can be added to support 802.15.4 on the 840DK
boards and it won't conflict with the bluetooth L2 selection.

NOTE: adding ENTROPY as a default selection for nRF52 HW, since
it was being selected by NET_BT prior to this patch.

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>