scripts/hawkbit.py: update logic to fix compatibility
In newer versions of HawkBit, the response from the post sending
the softwaremodules, does not include several of the key URLs
(artifacts, type, metadata).

Instead, we must query the newly created software module using
the "self" url and read those details.

These script changes were tested with the the existing 0.2.0M4
version, and the latest 0.2.1 branch.

Signed-off-by: Michael Scott <mike@foundries.io>