Set default logging level to 0
If left as the default of 3, then all LOG_* statements are kept in
the code instead of #ifdef'd out as they were previously when using
SYS_LOG_*.  We can evaluate this change more as time permits to see
if we can perhaps only keep the LOG_ERR() messages.

This reduces the flash usage back down to "normal" size post-Logger
integration into the Bluetooth and other stacks.

Signed-off-by: Michael Scott <mike@foundries.io>